consistency_check/requirements.txt
connexion[swagger-ui]
httpx
pyyaml
uvicorn
starlette